In DataAccess.js I commented out most of the quoteWatch function and all of parseQuotes since the search part is already taking place in the addQuote function.
In index.js I modified addQuote to use the new format. I also modified the WEBSOCKET.onmessage function in getSymbolData to get the symbol string from the event data instead of the symbol that’s passed in. This is because if there is more than one symbol in the users symbol list, then the ‘symbol’ being passed in was always the same and only one Quote in the list would have values for high, low, open, etc.